pyproject.toml
[project]
name = "hiho-pytorch-base"
version = "0.1.0"
authors = [
    { name = "Hiroshiba", email = "hihokaruta@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "torch>=2.7.1",
    "torch-optimizer>=0.3.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.2",
    "tensorboard>=2.19.0",
    "wandb>=0.21.0",
]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "W", "UP", "D", "D213", "PT"]
ignore = [
    "E501", # line-too-long
    "D400", # missing-trailing-period。日本語の「。」に対応していないため。
    "D403", # first-word-uncapitalized。日本語とは無関係であるため。日英混合時の挙動に一貫性が無いため。
]
unfixable = [
    "F401", # unused-import
    "F841", # unused-variable
]

[tool.ruff.lint.per-file-ignores]
"test/**/test_*.py" = ["D103"] # undocumented-public-function。pytest の対象 test_*() が D103 の対象になり、関数名から内容が自明でも docstring を強制されるため。

[tool.ruff.lint.isort]
known-first-party = ["hiho_pytorch_base"]
known-third-party = ["numpy", "torch"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.format]
docstring-code-format = true
